+++ src/SpreadSheet.Core/Commands/ICommand.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace SpreadSheet.Commands
+{
+    public interface ICommand
+    {
+        SpreadSheet Operate(string input,SpreadSheet spreadSheet = null);
+    }
+}