+++ src/SpreadSheet.Core/Commands/ICommandParser.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SpreadSheet.Commands
+{
+    public interface ICommandParser
+    {
+        ICommand PaserCommand(string input);
+    }
+}