+++ src/SpreadSheet.Core/Exceptions/SpreadSheetException.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace SpreadSheet.Exceptions
+{
+    public class SpreadSheetException : Exception
+    {
+        public SpreadSheetException(string exceptionMessage) : base(exceptionMessage)
+        {
+
+        }
+    }
+}