+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # 更新日志
 
 ## [Unreleased]
+### Add
+- 新增SpreadSheet指令的设计
+
+### Changed
+- 将修改SpreadSheetItem的名称修改为SpreadSheetCell